requirements.txt
sentence-transformers>=2.2.2
hdbscan>=0.8.33
scikit-learn>=1.1.3
pandas>=1.5.3
pyyaml>=6.0
streamlit>=1.22.0
networkx>=3.2.1
matplotlib>=3.7.1
